third_party/blink/web_tests/virtual/partitioned-popins-disabled/external/wpt/partitioned-popins/partitioned-popins.secure.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify only secure Partitioned Popins can be opened from secure page
  assert_equals: expected "Secure-PathOnly-" but got "Secure-Insecure-PathOnly-Empty-"
Harness: the test ran to completion.
